requirements.txt
alabaster==0.7.12
amqp==2.3.2
anaconda-client==1.7.2
anaconda-navigator==1.9.6
anaconda-project==0.8.2
argon2-cffi==18.3.0
arrow==0.12.1
asn1crypto==0.24.0
astroid==2.1.0
astropy==3.1
atomicwrites==1.2.1
attrs==18.2.0
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
billiard==3.5.0.5
binaryornot==0.4.4
bitarray==0.8.3
bkcharts==0.2
blaze==0.11.3
bleach==3.0.2
bokeh==1.0.2
boto==2.49.0
boto3==1.7.49
botocore==1.10.84
Bottleneck==1.2.1
cachetools==3.0.0
celery==4.2.1
certifi==2019.3.9
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
clyent==1.2.2
colorama==0.4.1
conda==4.6.8
conda-build==3.17.6
conda-verify==3.1.1
contextlib2==0.5.5
cookiecutter==1.6.0
coverage==4.5.2
cryptography==2.6.1
cssselect==1.0.3
cycler==0.10.0
Cython==0.29.2
cytoolz==0.9.0.1
dask==1.0.0
datashape==0.5.4
dateparser==0.7.0
decorator==4.3.0
defusedxml==0.5.0
distributed==1.25.1
Django==2.0.9
django-bulk-update==2.2.0
django-environ==0.4.5
django-model-utils==3.1.2
djangorestframework==3.9.2
docutils==0.14
easygui==0.98.1
EasyProcess==0.2.5
emoji==0.5.0
emot==2.1
entrypoints==0.2.3
et-xmlfile==1.0.1
fake-useragent==0.1.10
fastcache==1.0.2
feedfinder2==0.0.4
feedparser==5.2.1
filelock==3.0.10
Flask==1.0.2
Flask-Cors==3.0.7
flower==0.9.2
future==0.17.1
gevent==1.3.7
glob2==0.6
gmpy2==2.0.8
google-api-python-client==1.7.4
google-auth==1.6.2
google-auth-httplib2==0.0.3
greenlet==0.4.15
h5py==2.8.0
heapdict==1.0.0
html5lib==1.0.1
httplib2==0.11.3
idna==2.8
imageio==2.4.1
imagesize==1.1.0
importlib-metadata==0.6
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.4
jedi==0.13.2
jeepney==0.4
jieba3k==0.35.1
Jinja2==2.10
jinja2-time==0.2.0
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.3
jupyterlab-server==0.2.0
keyring==17.0.0
kiwisolver==1.0.1
kombu==4.2.2.post1
lazy-object-proxy==1.3.1
libarchive-c==2.8
lief==0.9.0
llvmlite==0.26.0
locket==0.2.0
lxml==4.2.5
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.2
more-itertools==4.3.0
mpmath==1.1.0
msgpack==0.5.6
multipledispatch==0.6.0
mypy==0.670
mypy-extensions==0.4.1
navigator-updater==0.2.1
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
newspaper3k==0.2.6
nltk==3.4
nose==1.3.7
notebook==5.7.4
numba==0.41.0
numexpr==2.6.8
numpy==1.15.4
numpydoc==0.8.0
oauth2client==4.0.0
oauthlib==2.1.0
odo==0.5.1
olefile==0.46
openpyxl==2.5.12
packaging==19.0
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
partd==0.3.9
path.py==11.5.0
pathlib2==2.3.3
patsy==0.5.1
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
piplapis-python==5.1.4
pkginfo==1.4.2
pluggy==0.8.0
ply==3.11
poyo==0.4.2
prometheus-client==0.5.0
prompt-toolkit==2.0.7
psutil==5.4.8
psycopg2==2.7.7
ptyprocess==0.6.0
py==1.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.2
pyflakes==2.0.0
Pygments==2.3.1
pylint==2.2.2
pyodbc==4.0.25
pyOpenSSL==19.0.0
pyparsing==2.3.0
PySocks==1.6.8
pytest==4.2.0
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.2.0
pytest-openfiles==0.3.1
pytest-remotedata==0.3.1
python-dateutil==2.7.5
python-slugify==2.0.1
python3-openid==3.1.0
pytz==2018.9
PyVirtualDisplay==0.2.1
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
QtAwesome==0.5.3
qtconsole==4.4.3
QtPy==1.5.2
redis==2.10.6
regex==2018.11.22
requests==2.21.0
requests-file==1.4.3
requests-oauthlib==1.0.0
rope==0.11.0
rsa==4.0
rson==0.9
ruamel-yaml==0.15.46
s3transfer==0.1.13
scikit-image==0.14.1
scikit-learn==0.20.1
scipy==1.1.0
seaborn==0.9.0
SecretStorage==3.1.0
selenium==3.12.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
slackclient==1.2.1
snowballstemmer==1.2.1
sortedcollections==1.0.1
sortedcontainers==2.1.0
Sphinx==1.8.4
sphinxcontrib-websupport==1.1.0
spyder==3.3.2
spyder-kernels==0.3.0
SQLAlchemy==1.2.15
statsmodels==0.9.0
sympy==1.3
tables==3.4.4
tblib==1.3.2
terminado==0.8.1
testpath==0.4.2
textblob==0.15.3
tldextract==2.2.0
toolz==0.9.0
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
tweepy==3.6.0
typed-ast==1.2.0
typing==3.6.4
tzlocal==1.5.1
unicodecsv==0.14.1
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.1
vine==1.2.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.54.0
Werkzeug==0.14.1
whichcraft==0.5.2
widgetsnbextension==3.4.2
wrapt==1.10.11
wurlitzer==1.0.2
xlrd==1.2.0
XlsxWriter==1.1.2
xlwt==1.3.0
zict==0.1.3
